requirements.txt
anthropic==0.49.0
pyyaml>=6.0

# Testing dependencies (install with: pip install -r requirements-test.txt)
# pytest>=7.0.0
# pytest-cov>=4.0.0  
# pytest-mock>=3.10.0
